requirements.txt
asgiref==3.3.1
astroid==2.3.3
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.1
certifi==2019.11.28
chardet==3.0.4
Click==7.0
colorama==0.4.3
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
decorator==4.4.1
Django==3.1.3
django-bootstrap4==1.0.1
django-cors-headers==3.5.0
django-debug-toolbar==2.2
django-filter==2.2.0
django-mdeditor==0.1.16
django-model-utils==3.2.0
django-rest-swagger==2.2.0
django-softdelete==0.9.2
django-uuslug==1.1.9
djangorestframework==3.12.2
djangorestframework-simplejwt==4.4.0
drf-yasg==1.20.0
gunicorn==20.0.4
idna==2.8
importlib-metadata==1.5.0
inflection==0.3.1
ipython==7.9.0
ipython-genutils==0.2.0
isort==4.3.21
itypes==1.1.0
jedi==0.15.1
Jinja2==2.10.3
jsonschema==3.2.0
lazy-object-proxy==1.4.3
Markdown==3.1.1
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
openapi-codec==1.3.2
packaging==20.3
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prompt-toolkit==2.0.10
psycopg2-binary==2.8.4
ptyprocess==0.6.0
Pygments==2.4.2
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.6
pyrsistent==0.15.7
python-slugify==4.0.0
pytz==2019.3
PyYAML==5.3
requests==2.22.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
Serializer==0.2.1
simplejson==3.17.0
six==1.13.0
soupsieve==1.9.5
sqlparse==0.3.0
swagger-spec-validator==2.5.0
text-unidecode==1.3
traitlets==4.3.3
typed-ast==1.4.1
uritemplate==3.0.0
urllib3==1.25.7
wcwidth==0.1.7
wrapt==1.11.2
zipp==3.1.0
